Feat : 친구 기능 추가 및 수정

+++ dondoc/backend/src/main/java/com/dondoc/backend/user/dto/user/FindUserDto.java
@@ -22,7 +22,7 @@ public class FindUserDto {
         private Long userId;
         private String phoneNumber;
         private String NickName;
-        private Long imageNumber;
+        private int imageNumber;
         private String msg;
         @JsonIgnore
         private boolean success;
